requirements.txt
streamlit == 0.51.0
pandas==0.25.1
json5==0.8.5
Pillow==7.1.2
firebase==3.0.1
